+++ CHANGELOG.md
@@ -11,6 +11,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Renamed existing `ptrace` wrappers to encourage namespacing ([#692](https://github.com/nix-rust/nix/pull/692))
+- Changed function signature of `socket()` and `socketpair()`. The `protocol` argument
+  has changed type from `c_int` to `SockProtocol`.
+  It accepts a `None` value for default protocol that was specified with zero using `c_int`.
+  ([#647](https://github.com/nix-rust/nix/pull/647))
 
 ## [0.9.0] 2017-07-23